python-translator-app/requirements.txt
grpcio
googletrans
